Simplify Roll20 code
