--- src/roll20/classes.js
@@ -1,2 +1,0 @@
-export const CLOSE_BUTTON = "vtt-close-button";
-export const NOTIFICATION = "vtt-notification";
